[update] ちょっとシンプルな実装方法をおもいついたので修正
